test: Refactor to add an e2e Installation interface
