+++ test/e2e/tests/custom_gwp_test.go
@@ -3,289 +3,14 @@
 package tests
 
 import (
-	"context"
-	"fmt"
-	"os"
 	"testing"
-	"time"
 
-	"github.com/onsi/gomega"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	corev1 "k8s.io/api/core/v1"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	gwv1 "sigs.k8s.io/gateway-api/apis/v1"
-
-	"github.com/kgateway-dev/kgateway/v2/pkg/kgateway/wellknown"
-	"github.com/kgateway-dev/kgateway/v2/pkg/utils/envutils"
-	"github.com/kgateway-dev/kgateway/v2/pkg/utils/helmutils"
-	"github.com/kgateway-dev/kgateway/v2/pkg/utils/kubeutils"
 	"github.com/kgateway-dev/kgateway/v2/test/e2e"
-	"github.com/kgateway-dev/kgateway/v2/test/e2e/testutils/helper"
-	"github.com/kgateway-dev/kgateway/v2/test/e2e/testutils/install"
-	"github.com/kgateway-dev/kgateway/v2/test/testutils"
+	"github.com/kgateway-dev/kgateway/v2/test/e2e/tests/customgwp"
 )
 
-var kgatewayGWP = `
-apiVersion: gateway.kgateway.dev/v1alpha1
-kind: GatewayParameters
-metadata:
-  name: custom-gwp
-  namespace: kgateway-test
-spec:
-  kube:
-    podTemplate:
-      extraLabels:
-        custom: custom-label
-`
-
-var kgatewayGWP2 = `
-apiVersion: gateway.kgateway.dev/v1alpha1
-kind: GatewayParameters
-metadata:
-  name: custom-gwp-2
-  namespace: kgateway-test
-spec:
-  kube:
-    podTemplate:
-      extraLabels:
-        another: label
-`
-
-var kgatewayGateway = `
-apiVersion: gateway.networking.k8s.io/v1
-kind: Gateway
-metadata:
-  name: gw
-spec:
-  gatewayClassName: kgateway
-  listeners:
-    - protocol: HTTP
-      port: 8080
-      name: http
-`
-
-const gatewayNamespace = "default"
-
-var proxyObjectMeta = metav1.ObjectMeta{
-	Name:      "gw",
-	Namespace: gatewayNamespace,
-}
-
-// TestCustomGWP tests that the helm chart's gatewayClassParametersRefs configures
-// the default GatewayClass parametersRef correctly.
-// The test installs CRDs, creates the custom GatewayParameters resource, installs kgateway,
-// verifies that the GatewayClass parametersRef is configured correctly, creates a Gateway,
-// and verifies that the gateway pod has the custom label defined in the GatewayParameters.
-// It then upgrades the helm chart to reference a different GatewayParameters resource,
-// verifies that the GatewayClass parametersRef is updated correctly,
-// and verifies that the gateway pod is still running (even though the ParametersRef has changed to non-existent resource).
-// It then creates the new GatewayParameters resource,
-// and verifies that the gateway pod has the new label defined in the new GatewayParameters resource.
+// TestCustomGWP tests that the helm chart's gatewayClassParametersRefs
+// configures the default GatewayClass parametersRef correctly.
 func TestCustomGWP(t *testing.T) {
-	ctx := t.Context()
-	installNs, nsEnvPredefined := envutils.LookupOrDefault(testutils.InstallNamespace, "kgateway-test")
-	rolloutStabilityValuesFile := e2e.ManifestPath("custom-gwp-rollout-stability.yaml")
-	testInstallation := e2e.CreateTestInstallation(
-		t,
-		&install.Context{
-			InstallNamespace:          installNs,
-			ProfileValuesManifestFile: e2e.CommonRecommendationManifest,
-			ValuesManifestFile:        e2e.ManifestPath("custom-gwp.yaml"),
-			ExtraHelmArgs:             []string{"--values", rolloutStabilityValuesFile},
-		},
-	)
-
-	// Set the env to the install namespace if it is not already set
-	if !nsEnvPredefined {
-		os.Setenv(testutils.InstallNamespace, installNs)
-	}
-
-	// We register the cleanup function before we actually perform the installation.
-	// This allows us to uninstall kgateway, in case the original installation only completed partially
-	testutils.Cleanup(t, func() {
-		ctx := context.Background() // when you have a custom Cleanup, you can't use t.Context within it as the context is canceled before t's cleanup is called
-		if !nsEnvPredefined {
-			os.Unsetenv(testutils.InstallNamespace)
-		}
-		if t.Failed() {
-			testInstallation.PreFailHandler(ctx, t)
-		}
-
-		testInstallation.UninstallKgateway(ctx, t)
-	})
-
-	// install CRDs for kgateway
-	testInstallation.InstallKgatewayCRDsFromLocalChart(ctx, t)
-
-	// Delete any kgateway-managed GatewayClass left over from a prior test run.
-	// The controller creates GatewayClasses via server-side apply, so `helm
-	// uninstall` from a prior test doesn't remove them; if a stale one exists
-	// with the wrong (e.g. nil) parametersRef, the assertions below race the
-	// reconciler and fail.
-	deleteStaleKgatewayGatewayClasses(ctx, t, testInstallation)
-
-	// create GatewayParameters for kgateway
-	err := testInstallation.Actions.Kubectl().Apply(ctx, []byte(kgatewayGWP))
-	if err != nil {
-		t.Fatalf("failed to create GatewayParameters: %v", err)
-	}
-
-	// install kgateway
-	testInstallation.InstallKgatewayFromLocalChart(ctx, t)
-
-	// Wait for GatewayClass to be created
-	testInstallation.AssertionsT(t).EventuallyObjectsExist(ctx, &gwv1.GatewayClass{
-		ObjectMeta: metav1.ObjectMeta{Name: wellknown.DefaultGatewayClassName},
-	})
-
-	// create Gateway
-	err = testInstallation.Actions.Kubectl().Apply(ctx, []byte(kgatewayGateway))
-	if err != nil {
-		t.Fatalf("failed to create Gateway: %v", err)
-	}
-
-	// Wait for the controller to reconcile the GatewayClass with the expected parametersRef.
-	// A stale GatewayClass may exist from a prior test run in the same cluster (it's created
-	// by the controller via SSA and is not removed by helm uninstall), so EventuallyObjectsExist
-	// above returns immediately without guaranteeing the controller has reconciled.
-	r := require.New(t)
-	expectedNamespace := gwv1.Namespace("kgateway-test")
-	r.EventuallyWithT(func(c *assert.CollectT) {
-		gc := &gwv1.GatewayClass{}
-		err := testInstallation.ClusterContext.Client.Get(ctx, client.ObjectKey{Name: wellknown.DefaultGatewayClassName}, gc)
-		require.NoError(c, err, "failed to get kgateway GatewayClass")
-		require.NotNil(c, gc.Spec.ParametersRef, "kgateway GatewayClass spec.parametersRef is nil")
-		assert.Equal(c, "custom-gwp", gc.Spec.ParametersRef.Name, "expected kgateway GatewayClass parametersRef.name to be 'custom-gwp'")
-		require.NotNil(c, gc.Spec.ParametersRef.Namespace, "kgateway GatewayClass spec.parametersRef.namespace is nil")
-		assert.Equal(c, expectedNamespace, *gc.Spec.ParametersRef.Namespace, "expected kgateway GatewayClass parametersRef.namespace to be '%s'", expectedNamespace)
-		// Verify kgateway GatewayClass uses GatewayParameters GVK (gateway.kgateway.dev)
-		assert.Equal(c, gwv1.Group("gateway.kgateway.dev"), gc.Spec.ParametersRef.Group, "expected kgateway GatewayClass parametersRef.group to be 'gateway.kgateway.dev'")
-		assert.Equal(c, gwv1.Kind("GatewayParameters"), gc.Spec.ParametersRef.Kind, "expected kgateway GatewayClass parametersRef.kind to be 'GatewayParameters'")
-	}, 60*time.Second, 200*time.Millisecond)
-
-	// Wait for Gateway to be accepted and deployment created
-	testInstallation.AssertionsT(t).EventuallyReadyReplicas(ctx, proxyObjectMeta, gomega.Equal(1))
-
-	// Verify the gateway pod has the custom label
-	verifyPodLabel(t, ctx, testInstallation, "custom", "custom-label", "")
-
-	// Upgrade Helm to reference different ParametersRef (kgatewayGWP2)
-	chartUri, err := helper.GetLocalChartPath(helmutils.ChartName, "")
-	if err != nil {
-		t.Fatalf("failed to get chart path: %v", err)
-	}
-	err = testInstallation.Actions.Helm().WithReceiver(os.Stdout).Upgrade(
-		ctx,
-		helmutils.InstallOpts{
-			Namespace:       installNs,
-			CreateNamespace: true,
-			ValuesFiles: []string{
-				e2e.CommonRecommendationManifest,
-				e2e.ManifestPath("custom-gwp-2.yaml"),
-				rolloutStabilityValuesFile,
-			},
-			ReleaseName: helmutils.ChartName,
-			ChartUri:    chartUri,
-			ExtraArgs:   []string{"--wait", "--timeout", "2m"},
-		})
-	if err != nil {
-		t.Fatalf("failed to upgrade Helm: %v", err)
-	}
-	testInstallation.AssertionsT(t).EventuallyKgatewayInstallSucceeded(ctx)
-
-	// Verify kgateway GatewayClass is updated with new ref
-	r.EventuallyWithT(func(c *assert.CollectT) {
-		gcUpdated := &gwv1.GatewayClass{}
-		err := testInstallation.ClusterContext.Client.Get(ctx, client.ObjectKey{Name: wellknown.DefaultGatewayClassName}, gcUpdated)
-		require.NoError(c, err, "failed to get kgateway GatewayClass after upgrade")
-		require.NotNil(c, gcUpdated.Spec.ParametersRef, "kgateway GatewayClass spec.parametersRef is nil after upgrade")
-		assert.Equal(c, "custom-gwp-2", gcUpdated.Spec.ParametersRef.Name, "expected kgateway GatewayClass parametersRef.name to be 'custom-gwp-2' after upgrade")
-		require.NotNil(c, gcUpdated.Spec.ParametersRef.Namespace, "kgateway GatewayClass spec.parametersRef.namespace is nil after upgrade")
-		assert.Equal(c, expectedNamespace, *gcUpdated.Spec.ParametersRef.Namespace, "expected kgateway GatewayClass parametersRef.namespace to be '%s' after upgrade", expectedNamespace)
-	}, 60*time.Second, 200*time.Millisecond)
-
-	// Ensure gateway pods are still running (even though the ParametersRef has changed to non-existent resource)
-	testInstallation.AssertionsT(t).EventuallyReadyReplicas(ctx, proxyObjectMeta, gomega.Equal(1))
-
-	// Create the kgatewayGWP2 GatewayParameters resource
-	err = testInstallation.Actions.Kubectl().Apply(ctx, []byte(kgatewayGWP2))
-	if err != nil {
-		t.Fatalf("failed to create GatewayParameters kgatewayGWP2: %v", err)
-	}
-
-	// Wait for Gateway to reconcile with new parameters
-	// The Gateway should pick up the new GatewayParameters and update the pods
-	testInstallation.AssertionsT(t).EventuallyReadyReplicas(ctx, proxyObjectMeta, gomega.Equal(1))
-
-	// Assert that eventually the deployment gateway pod is updated with the new label
-	r.EventuallyWithT(func(c *assert.CollectT) {
-		pods, err := kubeutils.GetReadyPodsForDeployment(ctx, testInstallation.ClusterContext.Clientset, proxyObjectMeta)
-		require.NoError(c, err, "failed to get ready pods for deployment after upgrade")
-		require.NotEmpty(c, pods, "no ready pods found for deployment after upgrade")
-
-		pod := &corev1.Pod{}
-		err = testInstallation.ClusterContext.Client.Get(ctx, client.ObjectKey{
-			Namespace: gatewayNamespace,
-			Name:      pods[0],
-		}, pod)
-		require.NoError(c, err, "failed to get pod after upgrade")
-		assert.NotNil(c, pod.Labels, "pod labels are nil after upgrade")
-		assert.Contains(c, pod.Labels, "another", "pod should have the 'another' label after upgrade")
-		assert.Equal(c, "label", pod.Labels["another"], "pod should have the new label 'another: label' after upgrade")
-	}, 60*time.Second, 2*time.Second)
-}
-
-// deleteStaleKgatewayGatewayClasses removes any GatewayClass managed by the
-// kgateway controller. Helm doesn't track these (the controller creates them
-// via server-side apply), so they survive `helm uninstall` and leak state
-// across tests.
-func deleteStaleKgatewayGatewayClasses(ctx context.Context, t *testing.T, testInstallation *e2e.TestInstallation) {
-	list := &gwv1.GatewayClassList{}
-	if err := testInstallation.ClusterContext.Client.List(ctx, list); err != nil {
-		t.Fatalf("failed to list GatewayClasses: %v", err)
-	}
-	for idx := range list.Items {
-		gc := &list.Items[idx]
-		if string(gc.Spec.ControllerName) != wellknown.DefaultGatewayControllerName {
-			continue
-		}
-		if err := testInstallation.ClusterContext.Client.Delete(ctx, gc); err != nil && !apierrors.IsNotFound(err) {
-			t.Fatalf("failed to delete stale GatewayClass %q: %v", gc.Name, err)
-		}
-	}
-}
-
-// verifyPodLabel checks that a pod for the given deployment eventually has the
-// specified label with the expected value. Uses EventuallyWithT to account for
-// the delay between pod creation and GatewayParameters reconciliation.
-func verifyPodLabel(
-	t *testing.T,
-	ctx context.Context,
-	testInstallation *e2e.TestInstallation,
-	labelKey string,
-	expectedValue string,
-	errorContext string,
-) {
-	require.New(t).EventuallyWithT(func(c *assert.CollectT) {
-		pods, err := kubeutils.GetReadyPodsForDeployment(ctx, testInstallation.ClusterContext.Clientset, proxyObjectMeta)
-		require.NoError(c, err, "failed to get ready pods for deployment %s", errorContext)
-		require.NotEmpty(c, pods, "no ready pods found for deployment %s", errorContext)
-
-		pod := &corev1.Pod{}
-		err = testInstallation.ClusterContext.Client.Get(ctx, client.ObjectKey{
-			Namespace: gatewayNamespace,
-			Name:      pods[0],
-		}, pod)
-		require.NoError(c, err, "failed to get pod %s", errorContext)
-
-		require.NotNil(c, pod.Labels, "pod labels are nil %s", errorContext)
-
-		labelValue, ok := pod.Labels[labelKey]
-		require.True(c, ok, fmt.Sprintf("pod does not have '%s' label %s", labelKey, errorContext))
-
-		assert.Equal(c, expectedValue, labelValue, "expected pod label '%s' to be '%s' %s", labelKey, expectedValue, errorContext)
-	}, 60*time.Second, 2*time.Second)
+	customgwp.Run(t, e2e.DefaultInstallationFactory)
 }
